changed capacitor recharge time to u64 cause it stored in miliseconds not seconds

+++ fitting_engine/src/stats/capacitor.rs
@@ -5,12 +5,12 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize, Stat, PartialEq)]
 pub struct Capacitor {
     pub capacitor_amount: f64,
-    pub capacitor_recharge_time: u16,
+    pub capacitor_recharge_time: u64,
     pub neut_resistance: f32,
 }
 
 impl Capacitor {
-    pub fn new(capacitor_amount: f64, capacitor_recharge_time: u16, neut_resistance: f32) -> Self {
+    pub fn new(capacitor_amount: f64, capacitor_recharge_time: u64, neut_resistance: f32) -> Self {
         Self {
             capacitor_amount,
             capacitor_recharge_time,
